requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
tensorflow==2.15.0
numpy==1.26.3
pandas==2.1.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
